// src/api/generated/models/ModelAreaDto.ts
/* generated using openapi-typescript-codegen -- do no edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */

import type { CoordinateDto } from './CoordinateDto';

export type ModelAreaDto = {
    modelAreaId: string;
    modelAreaType: string;
    coordinates: Array<CoordinateDto>;
};